+++ src/modules/progressIndicator.module.ts
@@ -1,8 +1,6 @@
-import {NgModule, ModuleWithProviders, FactoryProvider} from '@angular/core';
+import {NgModule} from '@angular/core';
 import {CommonModule as AngularCommonModule} from '@angular/common';
 import {ProgressIndicatorComponent} from '../components/progressIndicator/progressIndicator.component';
-import {ProgressIndicatorService} from '../components/progressIndicator/progressIndicator.service';
-import {ProgressIndicatorOptions} from '../components/progressIndicator/progressIndicatorOptions';
 
 /**
  * Module for progress indicator
@@ -15,39 +13,4 @@ import {ProgressIndicatorOptions} from '../components/progressIndicator/progress
 })
 export class ProgressIndicatorModule
 {
-    //######################### public methods #########################
-
-    /**
-     * Returns module with progress indicator service
-     */
-    public static forRoot(): ModuleWithProviders
-    {
-        return {
-            ngModule: ProgressIndicatorModule,
-            providers:
-            [
-                ProgressIndicatorService
-            ]
-        };
-    }
-
-    /**
-     * Returns module with progress indicator service with options
-     * @param options Factory method for options
-     */
-    public static forRootWithOptions(options: () => ProgressIndicatorOptions): ModuleWithProviders
-    {
-        return {
-            ngModule: ProgressIndicatorModule,
-            providers:
-            [
-                ProgressIndicatorService,
-                <FactoryProvider>
-                {
-                    provide: ProgressIndicatorOptions,
-                    useFactory: options
-                }
-            ]
-        };
-    }
 }